+++ src/main/java/com/ambev/techempowers/model/Produto.java
@@ -1,22 +1,14 @@
 package com.ambev.techempowers.model;
 
-public abstract class Produto {
-    String nome;
+import jakarta.persistence.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document(collection = "produtos")
+public class Produto {
+    @Id
+    private String id;
+    private String nome;
     private double preco;
 
-    public Produto(String nome, double preco) {
-        this.nome = nome;
-        this.preco = preco;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public double getPreco() {
-        return preco;
-    }
-
-    // Método abstrato para calcular o preço
-    public abstract double calcularPreco();
+    // getters e setters
 }
