+++ src/main/java/com/ambev/techempowers/service/ProdutoService.java
@@ -1,0 +1,22 @@
+package com.ambev.techempowers.service;
+
+import com.ambev.techempowers.model.Produto;
+import com.ambev.techempowers.repository.ProdutoRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProdutoService {
+    private final ProdutoRepository produtoRepository;
+
+    @Autowired
+    public ProdutoService(ProdutoRepository produtoRepository) {
+        this.produtoRepository = produtoRepository;
+    }
+
+    public Produto salvarProduto(Produto produto) {
+        return produtoRepository.save(produto);
+    }
+
+    // Outros métodos de serviço, como buscar, atualizar e excluir produtos
+}
