+++ src/main/resources/application.properties
@@ -12,3 +12,6 @@ spring.h2.console.path=/h2-console
 spring.application.name=tech-empowers
 server.port=8080
 
+spring.data.mongodb.uri=mongodb://localhost:27017/meumongo
+
+
